[#289] Refactor: enum & dto 위치 이동 및 네이밍 수정
