+++ src/main/java/umc/GrowIT/Server/converter/DiaryConverter.java
@@ -65,7 +65,7 @@ public class DiaryConverter {
                 .diaryId(diary.getId())
                 .content(diary.getContent())
                 .date(diary.getDate())
-                .creditInfo(CreditDTO.CreditInfo.builder()
+                .creditInfo(DiaryResponseDTO.CreditInfo.builder()
                         .granted(creditGranted)
                         .amount(creditGranted ? creditAmount : 0)
                         .build())
@@ -93,7 +93,7 @@ public class DiaryConverter {
                 .diaryId(diary.getId())
                 .content(diary.getContent())
                 .date(diary.getDate())
-                .creditInfo(CreditDTO.CreditInfo.builder()
+                .creditInfo(DiaryResponseDTO.CreditInfo.builder()
                         .granted(creditGranted)
                         .amount(creditGranted ? creditAmount : 0)
                         .build())
