+++ src/main/java/umc/GrowIT/Server/domain/enums/CreditTransactionType.java
@@ -1,0 +1,7 @@
+package umc.GrowIT.Server.domain.enums;
+
+public enum CreditTransactionType {
+    ALL,
+    EARN,
+    SPEND,
+}
