--- src/main/java/umc/GrowIT/Server/web/controller/enums/CreditFilterType.java
@@ -1,7 +1,0 @@
-package umc.GrowIT.Server.web.controller.enums;
-
-public enum CreditFilterType {
-    ALL,
-    EARN,
-    SPEND,
-}
